# api.yaml
openapi: 3.0.0
info:
  title: Store_Manage API
  description: API for managing a store's products, orders, and users.
  version: 1.0.0
...
paths:
  /auth/register:
    post:
      summary: Register a new user
      operationId: registerUser
      tags:
        - auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                  example: 1
                email:
                  type: string
                  example: user@example.com
                password:
                  type: string
                  example: Password123!
                repassword:
                  type: string
                  example: Password123!
      responses:
        200:
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User registered successfully
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      email:
                        type: string
                        example: user@example.com
                      isActive:
                        type: boolean
                        example: true
                      role:
                        type: string
                        example: customer
                      _id: 
                        type: string
                        example: 1234567890abcdef
                      createdAt: 
                        type: string
                        format: date-time
                        example: "2025-01-16T08:30:00Z"
                      updatedAt:
                        type: string
                        format: date-time
                        example: "2025-01-16T08:30:00Z"
        400:
          description: Bad request due to invalid data
  /auth/login:
    post:
      summary: Login a user
      operationId: loginUser
      tags:
        - auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: user@example.com
                password:
                  type: string
                  example: Password123!
      responses:
        200:
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  method:
                    type: string
                    example: LOGIN
                  data:
                    type: object
                    properties:
                      token:
                        type: string
                        example: abcdef1234567890
        400:
          description: Bad request due to invalid credentials
  /auth/logout:
    post:
      summary: Logout a user
      operationId: logoutUser
      security:
        - bearerAuth: []
      tags:
        - auth
      parameters:
        - in: header
          name: authorization
          required: true
          schema:
            type: string
            example: Bearer token_string
      responses:
        200:
          description: Successfully logged out
          content:
            application/json:
              schema:
                type: object
                properties:
                  method:
                    type: string
                    example: LOGOUT
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: Logged out successfully
        400:
          description: Bad request due to missing authorization token
  /auth/ban/{userId}:
    post:
      summary: Ban a user (Admin only)
      operationId: banUser
      security:
        - bearerAuth: []
      tags:
        - auth
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
            example: 1
        - in: header
          name: authorization
          required: true
          schema:
            type: string
            example: Bearer token_string
      responses:
        200:
          description: User banned successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User banned successfully
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      isActive:
                        type: boolean
                        example: false
        400:
          description: Bad request due to invalid user ID or role
  /auth/unban/{userId}:
    post:
      summary: Unban a user (Admin only)
      operationId: unbanUser
      security:
        - bearerAuth: []
      tags:
        - auth
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
            example: 1
        - in: header
          name: authorization
          required: true
          schema:
            type: string
            example: Bearer token_string
      responses:
        200:
          description: User unbanned successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User unbanned successfully
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      isActive:
                        type: boolean
                        example: true
        400:
          description: Bad request due to invalid user ID
  /auth/user-status/{userId}:
    get:
      summary: Get user status (Admin only)
      operationId: getUserStatus
      security:
        - bearerAuth: []
      tags:
        - auth
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
            example: 1
        - in: header
          name: authorization
          required: true
          schema:
            type: string
            example: Bearer token_string
      responses:
        200:
          description: Successfully retrieved user status
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      isActive:
                        type: boolean
                        example: true
        400:
          description: Bad request due to invalid user ID
  /auth/update/{id}:
    patch:
      summary: Update user details
      operationId: updateUser
      security:
        - bearerAuth: []
      tags:
        - auth
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            example: "1"
        - in: header
          name: authorization
          required: true
          schema:
            type: string
            example: Bearer token_string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: newuser@example.com
                password:
                  type: string
                  example: NewPassword123!
                repassword:
                  type: string
                  example: NewPassword123!
      responses:
        200:
          description: User details updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  method:
                    type: string
                    example: UPDATE
                  data:
                    type: object
                    properties:
                      email:
                        type: string
                        example: newuser@example.com
        400:
          description: Bad request due to invalid data or password mismatch
  /auth/delete/{id}:
    delete:
      summary: Delete a user (Admin only)
      operationId: deleteUser
      security:
        - bearerAuth: [] 
      tags:
        - auth
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            example: "1"
        - in: header
          name: authorization
          required: true
          schema:
            type: string
            example: Bearer token_string
      responses:
        200:
          description: User deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  method:
                    type: string
                    example: DELETE
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: User with ID 1 deleted successfully
        400:
          description: Bad request due to invalid user ID or unauthorized action
...